+++ src/data/loaders/zarr.ts
@@ -4,7 +4,7 @@ import * as zarrita from "zarrita";
 import {
   RawLabelsDataSource,
   createLabelsDataSource,
-} from "../../models/labels";
+} from "../../model/labels";
 import { UintArray } from "../../types";
 import { LabelsData } from "../labels";
 import { AbstractLabelsDataLoader } from "./base";
