+++ src/model/image.ts
